move parsing to its own file
